refactoring

+++ src/main/java/com/abalmas/dmytro/repository/AccountRepository.java
@@ -1,6 +1,6 @@
 package com.abalmas.dmytro.repository;
 
-import com.abalmas.dmytro.model.Entity.Account;
+import com.abalmas.dmytro.model.Account;
 import java.util.List;
 import java.util.Optional;
 
@@ -8,12 +8,12 @@ public interface AccountRepository {
 
   List<Account> findAll();
 
-  Optional<Account> findById(int id);
+  Optional <Account> findById(long id);
 
   Account add(Account account);
 
-  Account update(int id, Account account);
+  Account update(long id, Account account);
 
-  void delete(int id);
+  void delete(long id);
 
 }
